+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 * server: map FileNotFoundException and NoSuchFileException to 404 status code by default
 * core: TSIDGenerator.generateValue() can now accept desired timestamp instead of current one
+* jdbc: `PostgresNotifier.consumeNotifications()` will now reconnect automatically if connection is lost
 
 # 1.8.0
 * server: StatusCodeExceptions will now contain stack trace in dev mode